requirements.txt
Django==6.0
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
pillow==12.0.0
psycopg2-binary==2.9.11
celery==5.6.2
redis==6.4.0